move domain to sections
