+++ src/app/sections/domain/domain-info/domain-group/domain-group.component.scss
