+++ src/app/sections/domain/domain-info/domain-group/domain-group.module.ts
@@ -18,9 +18,9 @@ import { PipesModule } from '@vality/ng-core';
 
 import { PrettyJsonModule } from '@cc/components/pretty-json';
 
-import { TableModule } from '../../../../components/table';
-import { ThriftPipesModule } from '../../../shared';
-import { SelectModule } from '../../../shared/components/select';
+import { TableModule } from '../../../../../components/table';
+import { ThriftPipesModule } from '../../../../shared';
+import { SelectModule } from '../../../../shared/components/select';
 import { DomainGroupComponent } from './domain-group.component';
 
 @NgModule({
