+++ src/app/sections/domain/domain-info/domain-group/index.ts
