+++ src/app/sections/domain/domain-info/domain-group/types/data-source-item.ts
