+++ src/app/sections/domain/domain-info/domain-group/utils/filter-predicate.ts
