+++ src/app/sections/domain/domain-info/domain-info.component.scss
