+++ src/app/sections/domain/domain-info/domain-info.component.ts
@@ -9,12 +9,12 @@ import { filter, switchMap } from 'rxjs/operators';
 
 import { DomainMetadataViewExtensionsService } from '@cc/app/shared/services/domain-metadata-view-extensions';
 
-import { ConfirmActionDialogComponent } from '../../../components/confirm-action-dialog';
-import { enumHasValue } from '../../../utils';
-import { ViewerKind } from '../../shared/components/thrift-viewer';
-import { ErrorService } from '../../shared/services/error';
-import { NotificationService } from '../../shared/services/notification';
-import { DomainStoreService } from '../../thrift-services/damsel/domain-store.service';
+import { ConfirmActionDialogComponent } from '../../../../components/confirm-action-dialog';
+import { enumHasValue } from '../../../../utils';
+import { ViewerKind } from '../../../shared/components/thrift-viewer';
+import { ErrorService } from '../../../shared/services/error';
+import { NotificationService } from '../../../shared/services/notification';
+import { DomainStoreService } from '../../../thrift-services/damsel/domain-store.service';
 
 const VIEWER_KIND = 'domain-info-kind';
