+++ src/app/sections/domain/domain-info/domain-info.module.ts
@@ -11,10 +11,10 @@ import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { RouterModule } from '@angular/router';
 import { PipesModule, ActionsModule } from '@vality/ng-core';
 
-import { MonacoEditorModule } from '../../monaco-editor';
-import { ThriftPipesModule } from '../../shared';
-import { ThriftViewerModule } from '../../shared/components/thrift-viewer';
-import { DamselModule } from '../../thrift-services/damsel';
+import { MonacoEditorModule } from '../../../monaco-editor';
+import { ThriftPipesModule } from '../../../shared';
+import { ThriftViewerModule } from '../../../shared/components/thrift-viewer';
+import { DamselModule } from '../../../thrift-services/damsel';
 import { DomainGroupModule } from './domain-group';
 import { DomainInfoComponent } from './domain-info.component';
