+++ src/app/sections/domain/domain-info/index.ts
