+++ src/app/sections/domain/domain-obj-creation/domain-obj-creation.module.ts
@@ -4,34 +4,35 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
-import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatDialogModule } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { RouterModule } from '@angular/router';
 import { ActionsModule } from '@vality/ng-core';
 
-import { MonacoEditorModule } from '../../monaco-editor';
-import { ThriftEditorModule } from '../../shared/components/thrift-editor';
-import { ThriftViewerModule } from '../../shared/components/thrift-viewer';
-import { DomainObjReviewComponent } from './domain-obj-review.component';
+import { ThriftEditorModule } from '@cc/app/shared/components/thrift-editor';
+
+import { MonacoEditorModule } from '../../../monaco-editor';
+import { ThriftViewerModule } from '../../../shared/components/thrift-viewer';
+import { DomainObjCreationComponent } from './domain-obj-creation.component';
 
 @NgModule({
-    declarations: [DomainObjReviewComponent],
+    declarations: [DomainObjCreationComponent],
     imports: [
         CommonModule,
         FlexLayoutModule,
         RouterModule,
+        MatProgressSpinnerModule,
         MatCardModule,
         MatButtonModule,
-        MatCheckboxModule,
-        MonacoEditorModule,
         MatIconModule,
-        ThriftEditorModule,
-        MatProgressSpinnerModule,
-        ActionsModule,
+        MonacoEditorModule,
+        MatDialogModule,
         ReactiveFormsModule,
+        ThriftEditorModule,
+        ActionsModule,
         ThriftViewerModule,
     ],
-    exports: [DomainObjReviewComponent],
+    exports: [DomainObjCreationComponent],
 })
-export class DomainObjReviewModule {}
+export class DomainObjCreationModule {}
