+++ src/app/sections/domain/domain-obj-creation/index.ts
