+++ src/app/sections/domain/editor-container.scss
