+++ src/app/sections/domain/services/metadata.service.ts
@@ -4,7 +4,7 @@ import { Field } from '@vality/thrift-ts';
 import { from, Observable, of } from 'rxjs';
 import { map, shareReplay } from 'rxjs/operators';
 
-import { ThriftAstMetadata } from '../../api/utils';
+import { ThriftAstMetadata } from '../../../api/utils';
 
 @Injectable()
 export class MetadataService {
