+++ src/app/sections/domain/services/modified-domain-object.service.ts
