+++ src/app/sections/sections-routing.module.ts
@@ -3,6 +3,10 @@ import { RouterModule, Routes } from '@angular/router';
 
 const ROUTES: Routes = [
     {
+        path: 'domain',
+        loadChildren: () => import('./domain').then((m) => m.DomainModule),
+    },
+    {
         path: 'party',
         loadChildren: () => import('./party/party.module').then((m) => m.PartyModule),
     },
