+++ src/app/thrift-services/damsel/damsel.module.ts
@@ -1,12 +1,11 @@
 import { NgModule } from '@angular/core';
 
-import { DomainStoreService } from './domain-store.service';
 import { PaymentProcessingService } from './payment-processing.service';
 import { RoutingRulesModule } from './routing-rules';
 
 @NgModule({
     imports: [RoutingRulesModule],
-    providers: [PaymentProcessingService, DomainStoreService],
+    providers: [PaymentProcessingService],
 })
 /**
  * @deprecated
